Add screening for market cap

+++ API/Database/Model/User/CustomIndices/MarketCap.cs
@@ -2,7 +2,7 @@
 {
     public class MarketCap
     {
-        public decimal Upper;
-        public decimal Lower;
+        public double Upper;
+        public double Lower;
     }
 }
