+++ API/UserCustomIndices/Model/Response/MarketCap.cs
@@ -2,7 +2,7 @@
 {
     public struct MarketCap
     {
-        public decimal Upper;
-        public decimal Lower;
+        public double Upper;
+        public double Lower;
     }
 }
